+++ packages/atomic/src/components/atomic-no-results-v1/atomic-no-results.tsx
@@ -16,6 +16,7 @@ import {
   InitializeBindings,
 } from '../../utils/initialization-utils';
 import MagnifyingGlass from '../../images/magnifying-glass.svg';
+import escape from 'escape-html';
 
 /**
  * The `atomic-no-results` component displays search tips and a "Cancel last action" button when there are no results. Any additional content slotted inside of its element will be displayed as well.